requirements.txt
Flask==3.0.3
PyYAML==6.0.2
requests==2.32.3
colorama==0.4.6
